requirements.txt
Flask
requests
google-analytics-data
pytrends
gunicorn
# matplotlib